requirements.txt
openpyxl==3.0.10
PyQt6-Qt6==6.3.1
